passind data via props
