--- src/App.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import Post from './Post.jsx';
-import PostForm from './PostForm.jsx';
-
-class App extends React.Component {
-  render() {
-    return(
-      <div>
-        <Post />
-        <PostForm />
-      </div>
-    );
-  }
-}
-
-export default App;
