--- src/Post.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-class Post extends React.Component {
-  render() {
-    return <div>Супер новость</div>;
-  }
-}
-
-export default Post;
-
