+++ src/index.js
@@ -1,10 +1,8 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
-import App from './App.jsx';
+import App from './App.js';
 
 ReactDOM.render(
   <App />,
   document.getElementById('app')
 );
-
-
